requirements.txt
customtkinter==5.2.2
openai==1.65.0
python-docx==1.1.2
pdfplumber==0.11.4
